flask_project/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
altair==5.0.1
aniso8601==9.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
black==19.10b0
boto3==1.28.15
botocore==1.31.15
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
cdifflib==1.2.6
# certifi @ file:///croot/certifi_1690232220950/work/certifi
# cffi @ file:///tmp/build/80754af9/cffi_1636541934635/work
chardet==3.0.4
# charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.2
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
# cryptography @ file:///croot/cryptography_1689373676338/work
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7
decorator==5.1.1
diffusers==0.19.3
Distance==0.1.3
Django==4.2.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.20.1
editdistance==0.6.2
einops==0.6.1
exceptiongroup==1.1.2
executing==1.2.0
facenet-pytorch==2.5.3
faiss-cpu==1.7.4
fastapi==0.100.1
fasttext==0.9.2
ffmpy==0.3.1
filelock==3.12.2
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googletrans==3.0.0
gradio==3.28.3
gradio_client==0.3.0
grpcio==1.56.2
h11==0.9.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hsemotion==0.3.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.4
hydra-core==1.2.0
hyperframe==5.2.0
idna==2.10
ijson==3.2.3
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflect==7.0.0
iniconfig==2.0.0
ipadic==1.0.0
ipykernel==6.25.0
ipython==8.12.2
ipywidgets==8.0.7
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.2
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.4
kornia==0.6.12
latexcodec==2.0.1
lazy_loader==0.3
Levenshtein==0.21.1
librosa==0.10.0.post2
lightning-utilities==0.9.0
linkify-it-py==2.0.2
llvmlite==0.40.1
loguru==0.7.0
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mecab-python3==1.0.5
megatron-core==0.2.0
mkl-fft==1.3.1
# mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nemo-text-processing==0.1.8rc0
nemo-toolkit==1.19.1
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numba==0.57.1
# numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.14.0
OpenCC==1.1.6
opencv-python==4.8.0.74
orjson==3.9.2
packaging==23.1
pandas==2.0.3
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
plac==1.3.5
platformdirs==3.10.0
pluggy==1.2.0
pooch==1.6.0
portalocker==2.7.0
progress==1.6
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pybind11==2.11.1
pybtex==0.24.0
pybtex-docutils==1.0.2
# pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.15.1
pynini==2.1.5
# pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.0.9
pypinyin==0.49.0
pypinyin-dict==0.6.0
# PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.4.0
pytest-runner==6.0.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.1.0
rapidfuzz==2.13.7
referencing==0.30.0
regex==2023.6.3
# requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.5.0
rouge-score==0.1.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
sacrebleu==2.3.1
sacremoses==0.0.53
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.29.0
setproctitle==1.3.2
shellingham==1.5.0.post1
# six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
sox==1.4.1
soxr==0.3.5
Sphinx==7.1.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
text-unidecode==1.3
textdistance==4.5.0
texterrors==0.4.4
threadpoolctl==3.2.0
timm==0.6.5
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
torch-cluster==1.6.1+pt113cu117
torch-geometric==2.3.1
torchaudio==0.13.0
torchmetrics==1.0.1
torchvision==0.14.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typed-ast==1.5.5
typer==0.9.0
# typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
uc-micro-py==1.0.2
# urllib3 @ file:///croot/urllib3_1686163155763/work
uvicorn==0.23.2
wandb==0.15.7
wcwidth==0.2.6
webdataset==0.1.62
websockets==11.0.3
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.8
wrapt==1.15.0
yarl==1.9.2
youtokentome==1.0.6
zipp==3.16.2
